moved registration State
